requirements.txt
python-telegram-bot==21.6
anthropic==0.34.2
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.35
Pillow==10.4.0
exifread==3.0.0
python-dotenv==1.0.1
requests==2.32.3
APScheduler==3.10.4
Werkzeug==3.0.4
